% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_fmri_data.R
\name{delete_fmri_data}
\alias{delete_fmri_data}
\title{Delete fMRI Image Filenames}
\usage{
delete_fmri_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "fMRI"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the fMRI images
}
