% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fmri_data.R
\name{download_fmri_data}
\alias{download_fmri_data}
\title{Download fMRI Image Filenames}
\usage{
download_fmri_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "fMRI"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the fMRI images
}
