% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmri_filenames.R
\name{get_fmri_filenames}
\alias{get_fmri_filenames}
\title{Get fMRI Image Filenames}
\usage{
get_fmri_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "fMRI"} so it cannot be specified}
}
\value{
Vector of filenames
}
\description{
Return the filenames for the fMRI images
}
\examples{
get_fmri_filenames()
}
